requirements.txt
docx==0.2.4
dox==0.2.0
lxml==6.0.2
pbr==0.11.1
pillow==12.0.0
PyYAML==6.0.3
six==1.17.0
